move more test_dataframe tests
