+++ python/cudf/cudf/tests/test_dataframe.py
@@ -1,8 +1,6 @@
 # Copyright (c) 2018-2025, NVIDIA CORPORATION.
 
 import array as arr
-import datetime
-import decimal
 import io
 import operator
 import re
@@ -92,32 +90,6 @@ def _hide_concat_empty_dtype_warning():
             category=FutureWarning,
         )
         yield
-
-
-@pytest.fixture(
-    params=[
-        pd.DataFrame(
-            {
-                "a": [0, 1, 2, np.nan, 4, None, 6],
-                "b": [np.nan, None, "u", "h", "d", "a", "m"],
-            },
-            index=["q", "w", "e", "r", "t", "y", "u"],
-        ),
-        pd.DataFrame({"a": [0, 1, 2, 3, 4], "b": ["a", "b", "u", "h", "d"]}),
-        pd.DataFrame(
-            {
-                "a": [None, None, np.nan, None],
-                "b": [np.nan, None, np.nan, None],
-            }
-        ),
-        pd.DataFrame({"a": []}),
-        pd.DataFrame({"a": [np.nan], "b": [None]}),
-        pd.DataFrame({"a": ["a", "b", "c", None, "e"]}),
-        pd.DataFrame({"a": ["a", "b", "c", "d", "e"]}),
-    ]
-)
-def na_data(request):
-    return request.param
 
 
 @pytest.mark.parametrize("a", [[1, 2, 3], [1, 10, 30]])
@@ -5520,548 +5492,9 @@ def test_multiindex_wildcard_selection_partial():
     assert_eq(expect, got)
 
 
-@pytest_xfail(reason="Not yet properly supported.")
-def test_multiindex_wildcard_selection_three_level_all():
-    midx = cudf.MultiIndex.from_tuples(
-        [(c1, c2, c3) for c1 in "abcd" for c2 in "abc" for c3 in "ab"]
-    )
-    df = cudf.DataFrame({f"{i}": [i] for i in range(24)})
-    df.columns = midx
-
-    expect = df.to_pandas().loc[:, (slice("a", "c"), slice("a", "b"), "b")]
-    got = df.loc[:, (slice(None), "b")]
-    assert_eq(expect, got)
-
-
 def test_dataframe_assign_scalar_to_empty_series():
     expected = pd.DataFrame({"a": []})
     actual = cudf.DataFrame({"a": []})
     expected.a = 0
     actual.a = 0
     assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {0: [1, 2, 3], 2: [10, 11, 23]},
-        {("a", "b"): [1, 2, 3], ("2",): [10, 11, 23]},
-    ],
-)
-def test_non_string_column_name_to_arrow(data):
-    df = cudf.DataFrame(data)
-
-    expected = df.to_arrow()
-    actual = pa.Table.from_pandas(df.to_pandas())
-
-    assert expected.equals(actual)
-
-
-def test_complex_types_from_arrow():
-    expected = pa.Table.from_arrays(
-        [
-            pa.array([1, 2, 3]),
-            pa.array([10, 20, 30]),
-            pa.array([{"a": 9}, {"b": 10}, {"c": 11}]),
-            pa.array([[{"a": 1}], [{"b": 2}], [{"c": 3}]]),
-            pa.array([10, 11, 12]).cast(pa.decimal128(21, 2)),
-            pa.array([{"a": 9}, {"b": 10, "c": {"g": 43}}, {"c": {"a": 10}}]),
-        ],
-        names=["a", "b", "c", "d", "e", "f"],
-    )
-
-    df = cudf.DataFrame.from_arrow(expected)
-    actual = df.to_arrow()
-
-    assert expected.equals(actual)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {
-            "brand": ["Yum Yum", "Yum Yum", "Indomie", "Indomie", "Indomie"],
-            "style": ["cup", "cup", "cup", "pack", "pack"],
-            "rating": [4, 4, 3.5, 15, 5],
-        },
-        {
-            "brand": ["Indomie", "Yum Yum", "Indomie", "Indomie", "Indomie"],
-            "style": ["cup", "cup", "cup", "cup", "pack"],
-            "rating": [4, 4, 3.5, 4, 5],
-        },
-    ],
-)
-@pytest.mark.parametrize(
-    "subset", [None, ["brand"], ["rating"], ["style", "rating"]]
-)
-@pytest.mark.parametrize("keep", ["first", "last", False])
-def test_dataframe_duplicated(data, subset, keep):
-    gdf = cudf.DataFrame(data)
-    pdf = gdf.to_pandas()
-
-    expected = pdf.duplicated(subset=subset, keep=keep)
-    actual = gdf.duplicated(subset=subset, keep=keep)
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"col": [{"a": 1.1}, {"a": 2.1}, {"a": 10.0}, {"a": 11.2323}, None]},
-        {"a": [[{"b": 567}], None] * 10},
-        {"a": [decimal.Decimal(10), decimal.Decimal(20), None]},
-    ],
-)
-def test_dataframe_values_complex_types(data):
-    gdf = cudf.DataFrame(data)
-    with pytest.raises(NotImplementedError):
-        gdf.values
-
-
-def test_dataframe_from_arrow_slice():
-    table = pa.Table.from_pandas(
-        pd.DataFrame.from_dict(
-            {"a": ["aa", "bb", "cc"] * 3, "b": [1, 2, 3] * 3}
-        )
-    )
-    table_slice = table.slice(3, 7)
-
-    expected = table_slice.to_pandas()
-    actual = cudf.DataFrame.from_arrow(table_slice)
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "data,index",
-    [
-        ({"a": [1, 2, 3], "b": ["x", "y", "z", "z"], "c": 4}, None),
-        (
-            {
-                "a": [1, 2, 3],
-                "b": ["x", "y", "z"],
-            },
-            [10, 11],
-        ),
-        (
-            {
-                "a": [1, 2, 3],
-                "b": ["x", "y", "z"],
-            },
-            [10, 11],
-        ),
-        ([[10, 11], [12, 13]], ["a", "b", "c"]),
-    ],
-)
-def test_dataframe_init_length_error(data, index):
-    assert_exceptions_equal(
-        lfunc=pd.DataFrame,
-        rfunc=cudf.DataFrame,
-        lfunc_args_and_kwargs=(
-            [],
-            {"data": data, "index": index},
-        ),
-        rfunc_args_and_kwargs=(
-            [],
-            {"data": data, "index": index},
-        ),
-    )
-
-
-def test_dataframe_binop_with_mixed_date_types():
-    rng = np.random.default_rng(seed=0)
-    df = pd.DataFrame(
-        rng.random(size=(2, 2)),
-        columns=pd.Index(["2000-01-03", "2000-01-04"], dtype="datetime64[ns]"),
-    )
-    ser = pd.Series(rng.random(size=3), index=[0, 1, 2])
-    gdf = cudf.from_pandas(df)
-    gser = cudf.from_pandas(ser)
-    expected = df - ser
-    got = gdf - gser
-    assert_eq(expected, got)
-
-
-def test_dataframe_binop_with_mixed_string_types():
-    rng = np.random.default_rng(seed=0)
-    df1 = pd.DataFrame(rng.random(size=(3, 3)), columns=pd.Index([0, 1, 2]))
-    df2 = pd.DataFrame(
-        rng.random(size=(6, 6)),
-        columns=pd.Index([0, 1, 2, "VhDoHxRaqt", "X0NNHBIPfA", "5FbhPtS0D1"]),
-    )
-    gdf1 = cudf.from_pandas(df1)
-    gdf2 = cudf.from_pandas(df2)
-
-    expected = df2 + df1
-    got = gdf2 + gdf1
-
-    assert_eq(expected, got)
-
-
-def test_dataframe_binop_and_where():
-    rng = np.random.default_rng(seed=0)
-    df = pd.DataFrame(rng.random(size=(2, 2)), columns=pd.Index([True, False]))
-    gdf = cudf.from_pandas(df)
-
-    expected = df > 1
-    got = gdf > 1
-
-    assert_eq(expected, got)
-
-    expected = df[df > 1]
-    got = gdf[gdf > 1]
-
-    assert_eq(expected, got)
-
-
-def test_dataframe_binop_with_datetime_index():
-    rng = np.random.default_rng(seed=0)
-    df = pd.DataFrame(
-        rng.random(size=(2, 2)),
-        columns=pd.Index(["2000-01-03", "2000-01-04"], dtype="datetime64[ns]"),
-    )
-    ser = pd.Series(
-        rng.random(2),
-        index=pd.Index(
-            [
-                "2000-01-04",
-                "2000-01-03",
-            ],
-            dtype="datetime64[ns]",
-        ),
-    )
-    gdf = cudf.from_pandas(df)
-    gser = cudf.from_pandas(ser)
-    expected = df - ser
-    got = gdf - gser
-    assert_eq(expected, got)
-
-
-@pytest.mark.parametrize(
-    "dtype", ["datetime64[ns]", "timedelta64[ns]", "int64", "float32"]
-)
-def test_dataframe_mixed_dtype_error(dtype):
-    pdf = pd.Series([1, 2, 3], dtype=dtype).to_frame().astype(object)
-    with pytest.raises(TypeError):
-        cudf.from_pandas(pdf)
-
-
-@pytest.mark.parametrize(
-    "index_data,name",
-    [([10, 13], "a"), ([30, 40, 20], "b"), (["ef"], "c"), ([2, 3], "Z")],
-)
-def test_dataframe_reindex_with_index_names(index_data, name):
-    gdf = cudf.DataFrame(
-        {
-            "a": [10, 12, 13],
-            "b": [20, 30, 40],
-            "c": cudf.Series(["ab", "cd", "ef"], dtype="category"),
-        }
-    )
-    if name in gdf.columns:
-        gdf = gdf.set_index(name)
-    pdf = gdf.to_pandas()
-
-    gidx = cudf.Index(index_data, name=name)
-    actual = gdf.reindex(gidx)
-    expected = pdf.reindex(gidx.to_pandas())
-
-    assert_eq(actual, expected)
-
-    actual = gdf.reindex(index_data)
-    expected = pdf.reindex(index_data)
-
-    assert_eq(actual, expected)
-
-
-def test_dataframe_reduction_error():
-    gdf = cudf.DataFrame(
-        {
-            "a": cudf.Series([1, 2, 3], dtype="float"),
-            "d": cudf.Series([10, 20, 30], dtype="timedelta64[ns]"),
-        }
-    )
-
-    with pytest.raises(TypeError):
-        gdf.sum()
-
-
-def test_dataframe_from_generator():
-    pdf = pd.DataFrame((i for i in range(5)))
-    gdf = cudf.DataFrame((i for i in range(5)))
-    assert_eq(pdf, gdf)
-
-
-@pytest.mark.parametrize("name", ["a", 0, None, np.nan, cudf.NA])
-@pytest.mark.parametrize("contains", ["a", 0, None, np.nan, cudf.NA])
-@pytest.mark.parametrize("other_names", [[], ["b", "c"], [1, 2]])
-def test_dataframe_contains(name, contains, other_names):
-    column_names = [name, *other_names]
-    gdf = cudf.DataFrame({c: [0] for c in column_names})
-    pdf = pd.DataFrame({c: [0] for c in column_names})
-
-    assert_eq(gdf, pdf)
-
-    if contains is cudf.NA or name is cudf.NA:
-        expectation = contains is cudf.NA and name is cudf.NA
-        assert (contains in pdf) == expectation
-        assert (contains in gdf) == expectation
-    elif gdf.columns.dtype.kind == "f":
-        # In some cases, the columns are converted to an Index[float] based on
-        # the other column names. That casts name values from None to np.nan.
-        expectation = contains is np.nan and (name is None or name is np.nan)
-        assert (contains in pdf) == expectation
-        assert (contains in gdf) == expectation
-    else:
-        expectation = contains == name or (
-            contains is np.nan and name is np.nan
-        )
-        assert (contains in pdf) == expectation
-        assert (contains in gdf) == expectation
-
-    assert (contains in pdf) == (contains in gdf)
-
-
-def test_dataframe_series_dot():
-    pser = pd.Series(range(2))
-    gser = cudf.from_pandas(pser)
-
-    expected = pser @ pser
-    actual = gser @ gser
-
-    assert_eq(expected, actual)
-
-    pdf = pd.DataFrame([[1, 2], [3, 4]], columns=list("ab"))
-    gdf = cudf.from_pandas(pdf)
-
-    expected = pser @ pdf
-    actual = gser @ gdf
-
-    assert_eq(expected, actual)
-
-    assert_exceptions_equal(
-        lfunc=pdf.dot,
-        rfunc=gdf.dot,
-        lfunc_args_and_kwargs=([pser], {}),
-        rfunc_args_and_kwargs=([gser], {}),
-    )
-
-    assert_exceptions_equal(
-        lfunc=pdf.dot,
-        rfunc=gdf.dot,
-        lfunc_args_and_kwargs=([pdf], {}),
-        rfunc_args_and_kwargs=([gdf], {}),
-    )
-
-    pser = pd.Series(range(2), index=["a", "k"])
-    gser = cudf.from_pandas(pser)
-
-    pdf = pd.DataFrame([[1, 2], [3, 4]], columns=list("ab"), index=["a", "k"])
-    gdf = cudf.from_pandas(pdf)
-
-    expected = pser @ pdf
-    actual = gser @ gdf
-
-    assert_eq(expected, actual)
-
-    actual = gdf @ [2, 3]
-    expected = pdf @ [2, 3]
-
-    assert_eq(expected, actual)
-
-    actual = pser @ [12, 13]
-    expected = gser @ [12, 13]
-
-    assert_eq(expected, actual)
-
-
-def test_dataframe_reindex_keep_colname():
-    gdf = cudf.DataFrame([1], columns=cudf.Index([1], name="foo"))
-    result = gdf.reindex(index=[0, 1])
-    expected = cudf.DataFrame(
-        [1, None], columns=cudf.Index([1], name="foo"), index=[0, 1]
-    )
-    assert_eq(result, expected)
-
-
-def test_dataframe_duplicate_index_reindex():
-    gdf = cudf.DataFrame({"a": [0, 1, 2, 3]}, index=[0, 0, 1, 1])
-    pdf = gdf.to_pandas()
-
-    assert_exceptions_equal(
-        gdf.reindex,
-        pdf.reindex,
-        lfunc_args_and_kwargs=([10, 11, 12, 13], {}),
-        rfunc_args_and_kwargs=([10, 11, 12, 13], {}),
-    )
-
-
-@pytest.mark.parametrize(
-    "expected",
-    [
-        pd.RangeIndex(1, 2, name="a"),
-        pd.Index([1], dtype=np.int8, name="a"),
-        pd.MultiIndex.from_arrays([[1]], names=["a"]),
-    ],
-)
-@pytest.mark.parametrize("binop", [lambda df: df == df, lambda df: df - 1])
-def test_dataframe_binop_preserves_column_metadata(expected, binop):
-    df = cudf.DataFrame([1], columns=expected)
-    result = binop(df).columns
-    pd.testing.assert_index_equal(result, expected, exact=True)
-
-
-@pytest.mark.parametrize(
-    "scalar",
-    [
-        1,
-        1.0,
-        "a",
-        datetime.datetime(2020, 1, 1),
-        datetime.timedelta(1),
-        {"1": 2},
-        [1],
-        decimal.Decimal("1.0"),
-    ],
-)
-def test_dataframe_to_pandas_arrow_type_nullable_raises(scalar):
-    pa_array = pa.array([scalar, None])
-    df = cudf.DataFrame({"a": pa_array})
-    with pytest.raises(ValueError):
-        df.to_pandas(nullable=True, arrow_type=True)
-
-
-@pytest.mark.parametrize(
-    "scalar",
-    [
-        1,
-        1.0,
-        "a",
-        datetime.datetime(2020, 1, 1),
-        datetime.timedelta(1),
-        {"1": 2},
-        [1],
-        decimal.Decimal("1.0"),
-    ],
-)
-def test_dataframe_to_pandas_arrow_type(scalar):
-    pa_array = pa.array([scalar, None])
-    df = cudf.DataFrame({"a": pa_array})
-    result = df.to_pandas(arrow_type=True)
-    expected = pd.DataFrame({"a": pd.arrays.ArrowExtensionArray(pa_array)})
-    pd.testing.assert_frame_equal(result, expected)
-
-
-@pytest.mark.parametrize("dtype1", ["int16", "float32"])
-@pytest.mark.parametrize("dtype2", ["int16", "float32"])
-def test_dataframe_loc_int_float(dtype1, dtype2):
-    df = cudf.DataFrame(
-        {"a": [10, 11, 12, 13, 14]},
-        index=cudf.Index([1, 2, 3, 4, 5], dtype=dtype1),
-    )
-    pdf = df.to_pandas()
-
-    gidx = cudf.Index([2, 3, 4], dtype=dtype2)
-    pidx = gidx.to_pandas()
-
-    actual = df.loc[gidx]
-    expected = pdf.loc[pidx]
-
-    assert_eq(actual, expected, check_index_type=True, check_dtype=True)
-
-
-@pytest.mark.parametrize("names", [["abc", "def"], [1, 2], ["abc", 10]])
-def test_dataframe_multiindex_column_names(names):
-    arrays = [["A", "A", "B", "B"], ["one", "two", "one", "two"]]
-    tuples = list(zip(*arrays, strict=True))
-    index = pd.MultiIndex.from_tuples(tuples, names=["first", "second"])
-
-    pdf = pd.DataFrame([[1, 2, 3, 4], [5, 6, 7, 8]], columns=index)
-    df = cudf.from_pandas(pdf)
-
-    assert_eq(df, pdf)
-    assert_eq(df.columns.names, pdf.columns.names)
-    pdf.columns.names = names
-    df.columns.names = names
-    assert_eq(df, pdf)
-    assert_eq(df.columns.names, pdf.columns.names)
-
-
-def test_roundtrip_dataframe_plc_table(na_data):
-    pdf = na_data
-    expect = cudf.DataFrame.from_pandas(pdf)
-    actual = cudf.DataFrame.from_pylibcudf(*expect.to_pylibcudf())
-    assert_eq(expect, actual)
-
-
-def test_dataframe_midx_columns_loc():
-    idx_1 = ["Hi", "Lo"]
-    idx_2 = ["I", "II", "III"]
-    idx = cudf.MultiIndex.from_product([idx_1, idx_2])
-
-    data_rand = (
-        np.random.default_rng(seed=0)
-        .uniform(0, 1, 3 * len(idx))
-        .reshape(3, -1)
-    )
-    df = cudf.DataFrame(data_rand, index=["A", "B", "C"], columns=idx)
-    pdf = df.to_pandas()
-
-    assert_eq(df.shape, pdf.shape)
-
-    expected = pdf.loc[["A", "B"]]
-    actual = df.loc[["A", "B"]]
-
-    assert_eq(expected, actual)
-    assert_eq(df, pdf)
-
-
-def test_rename_reset_label_dtype():
-    data = {1: [2]}
-    col_mapping = {1: "a"}
-    result = cudf.DataFrame(data).rename(columns=col_mapping)
-    expected = pd.DataFrame(data).rename(columns=col_mapping)
-    assert_eq(result, expected)
-
-
-def test_insert_reset_label_dtype():
-    result = cudf.DataFrame({1: [2]})
-    expected = pd.DataFrame({1: [2]})
-    result.insert(1, "a", [2])
-    expected.insert(1, "a", [2])
-    assert_eq(result, expected)
-
-
-def test_setitem_reset_label_dtype():
-    result = cudf.DataFrame({1: [2]})
-    expected = pd.DataFrame({1: [2]})
-    result["a"] = [2]
-    expected["a"] = [2]
-    assert_eq(result, expected)
-
-
-def test_dataframe_midx_cols_getitem():
-    df = cudf.DataFrame(
-        {
-            "a": ["a", "b", "c"],
-            "b": ["b", "", ""],
-            "c": [10, 11, 12],
-        }
-    )
-    df.columns = df.set_index(["a", "b"]).index
-    pdf = df.to_pandas()
-
-    expected = df["c"]
-    actual = pdf["c"]
-    assert_eq(expected, actual)
-    df = cudf.DataFrame(
-        [[1, 0], [0, 1]],
-        columns=[
-            ["foo", "foo"],
-            ["location", "location"],
-            ["x", "y"],
-        ],
-    )
-    df = df.assign(bools=cudf.Series([True, False], dtype="bool"))
-    assert_eq(df["bools"], df.to_pandas()["bools"])
